+++ XBMC Remote/Utilities.m
@@ -56,7 +56,7 @@
 }
 
 - (UIColor *)slightLighterColorForColor:(UIColor *)c{
-    float r, g, b, a;
+    CGFloat r, g, b, a;
     if ([c getRed:&r green:&g blue:&b alpha:&a])
         return [UIColor colorWithRed:MIN(r + 0.2, 1.0)
                                green:MIN(g + 0.2, 1.0)
@@ -66,7 +66,7 @@
 }
 
 - (UIColor *)lighterColorForColor:(UIColor *)c{
-    float r, g, b, a;
+    CGFloat r, g, b, a;
     if ([c getRed:&r green:&g blue:&b alpha:&a])
         return [UIColor colorWithRed:MIN(r + 0.4, 1.0)
                                green:MIN(g + 0.4, 1.0)
@@ -76,7 +76,7 @@
 }
 
 - (UIColor *)darkerColorForColor:(UIColor *)c{
-    float r, g, b, a;
+    CGFloat r, g, b, a;
     if ([c getRed:&r green:&g blue:&b alpha:&a])
         return [UIColor colorWithRed:MAX(r - 0.1, 0.0)
                                green:MAX(g - 0.1, 0.0)
